correct compile error
Former-commit-id: 99fbc8db6dc18651dbfae30c4672d86de29961d9

+++ src/ves/xplorer/minerva/NavigateToModel.cxx
@@ -83,7 +83,7 @@ void NavigateToModel::Execute ( CommandPtr command, MinervaManager& manager )
 
   osg::BoundingSphere boundingSphere ( dcs->getBound() );
   osg::Vec3d location ( modelWrapper->location() );
-  const double altitudeOffset ( gmtl::Math::Max ( boundingSphere.radius() * 2, 30000.0 ) );
+  const double altitudeOffset ( gmtl::Math::Max( double( boundingSphere.radius() * 2 ), 30000.0 ) );
 
   Minerva::Core::Data::Camera::RefPtr camera ( new Minerva::Core::Data::Camera );
   camera->longitude ( location[0] );
